@@ -1,32 +1,32 @@
 {
-  "name": "nextjs-tailwind-portfolio-page",
-  "version": "1.0.0",
+  "name": "app-gerenciamento-financeiro-frontend",
+  "version": "0.1.0",
   "private": true,
-  "author": "Creative Tim",
-  "license": "See license in https://www.creative-tim.com/license",
-  "homepage": "https://demos.creative-tim.com/nextjs-tailwind-portfolio-page",
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev --port 9080",
     "build": "next build",
     "start": "next start",
     "lint": "next lint"
   },
   "dependencies": {
-    "@heroicons/react": "2.0.18",
-    "@material-tailwind/react": "2.1.2",
-    "next": "13.4.0",
-    "react": "18",
-    "react-dom": "18"
+    "@headlessui/react": "^1.7.17",
+    "@heroicons/react": "^2.0.18",
+    "axios": "^1.6.2",
+    "next": "14.0.3",
+    "react": "^18",
+    "react-dom": "^18",
+    "react-hook-form": "^7.48.2",
+    "tw-elements": "^1.1.0"
   },
   "devDependencies": {
-    "@types/node": "20",
-    "@types/react": "18",
-    "@types/react-dom": "18",
-    "autoprefixer": "10",
-    "eslint": "8",
-    "eslint-config-next": "13.5.4",
-    "postcss": "8",
-    "tailwindcss": "3",
-    "typescript": "5"
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
+    "autoprefixer": "^10.4.16",
+    "eslint": "^8",
+    "eslint-config-next": "14.0.3",
+    "postcss": "^8.4.32",
+    "tailwindcss": "^3.3.6",
+    "typescript": "^5"
   }
 }
